feat(hardhat): create arena lobby from base game
Tested out the upgrade and deploy flow for creating new versions of arena mode
